+++ src/pages/Hybrid Evaluation/Start.tsx
@@ -14,11 +14,6 @@ import {
   type CreditEvaluationCreateRequest,
   type CreditEvaluationResponse,
 } from '../../api/creditEvaluationApi';
-import { 
-  storeSummaryApi, 
-  type StoreSummaryCsvUploadRequest,
-  type SalesDataRow 
-} from '../../api/storeSummaryApi';
 import { useAuthStore } from '../../stores/useAuthStore';
 import { EsgSection } from './components/EsgSection';
 import { SalesSection } from './components/SalesSection';
@@ -208,178 +203,6 @@ const StartHybridEvaluation = () => {
     useState<CreditEvaluationResponse | null>(null);
   const [error, setError] = useState<string | null>(null);
   
-  // CSV 업로드 관련 상태
-  const [isUploadingCsv, setIsUploadingCsv] = useState(false);
-  const [csvUploadError, setCsvUploadError] = useState<string | null>(null);
-  const [csvUploadSuccess, setCsvUploadSuccess] = useState(false);
-
-  // CSV 파일 파싱 함수
-  const parseCsvFile = (file: File): Promise<SalesDataRow[]> => {
-    return new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        try {
-          const text = e.target?.result as string;
-          const lines = text.split('\n').filter(line => line.trim());
-          
-          if (lines.length < 2) {
-            reject(new Error('CSV 파일이 비어있거나 헤더만 있습니다.'));
-            return;
-          }
-          
-          // 첫 번째 줄은 헤더
-          const headers = lines[0].split(',').map(h => h.trim());
-          const expectedHeaders = [
-            'total_sales_amount', 'weekday_sales_amount', 'weekend_sales_amount',
-            'lunch_sales_ratio', 'dinner_sales_ratio', 'transaction_count',
-            'weekday_transaction_count', 'weekend_transaction_count', 'mom_growth_rate',
-            'yoy_growth_rate', 'sales_cv', 'avg_transaction_value',
-            'cash_payment_ratio', 'card_payment_ratio', 'revisit_customer_sales_ratio',
-            'new_customer_ratio'
-          ];
-          
-          // 헤더 검증
-          const missingHeaders = expectedHeaders.filter(header => !headers.includes(header));
-          if (missingHeaders.length > 0) {
-            reject(new Error(`필수 컬럼이 누락되었습니다: ${missingHeaders.join(', ')}`));
-            return;
-          }
-          
-          // 데이터 파싱
-          const salesData: SalesDataRow[] = [];
-          for (let i = 1; i < lines.length; i++) {
-            const values = lines[i].split(',').map(v => v.trim());
-            if (values.length !== headers.length) {
-              reject(new Error(`${i + 1}번째 행의 컬럼 수가 일치하지 않습니다.`));
-              return;
-            }
-            
-            const row: SalesDataRow = {};
-            headers.forEach((header, index) => {
-              const value = values[index];
-              if (value && value !== '') {
-                // snake_case를 camelCase로 변환
-                const camelCaseHeader = header.replace(/_([a-z])/g, (match, letter) => letter.toUpperCase());
-                (row as any)[camelCaseHeader] = parseFloat(value) || 0;
-                console.debug(`🔄 변환: ${header} -> ${camelCaseHeader} = ${value}`);
-              }
-            });
-            salesData.push(row);
-          }
-          
-          if (salesData.length === 0) {
-            reject(new Error('CSV 파일에 데이터가 없습니다.'));
-            return;
-          }
-          
-          resolve(salesData);
-        } catch (error) {
-          reject(new Error(`CSV 파싱 중 오류가 발생했습니다: ${error instanceof Error ? error.message : error}`));
-        }
-      };
-      reader.onerror = () => reject(new Error('파일을 읽을 수 없습니다.'));
-      reader.readAsText(file);
-    });
-  };
-
-  // CSV 파일 업로드 처리 함수
-  const handleCsvUpload = async (file: File) => {
-    if (!sessionId) {
-      setCsvUploadError('세션 ID가 없습니다.');
-      return;
-    }
-
-    // 파일 유효성 검사
-    if (!file.name.toLowerCase().endsWith('.csv')) {
-      setCsvUploadError('CSV 파일만 업로드 가능합니다.');
-      return;
-    }
-
-    if (file.size > 10 * 1024 * 1024) { // 10MB 제한
-      setCsvUploadError('파일 크기가 10MB를 초과할 수 없습니다.');
-      return;
-    }
-
-    if (file.size === 0) {
-      setCsvUploadError('빈 파일은 업로드할 수 없습니다.');
-      return;
-    }
-
-    try {
-      setIsUploadingCsv(true);
-      setCsvUploadError(null);
-      setCsvUploadSuccess(false);
-      
-      console.log('🔄 CSV 파일 파싱 시작:', file.name);
-      
-      // CSV 파일 파싱
-      const salesData = await parseCsvFile(file);
-      console.log('📊 파싱된 데이터:', salesData);
-      
-      // API 요청 데이터 구성
-      const uploadRequest: StoreSummaryCsvUploadRequest = {
-        sessionId: parseInt(sessionId),
-        businessRegistrationNo: '000-00-00000', // 기본값 또는 사용자 입력값
-        salesData: salesData
-      };
-      
-      console.log('📡 CSV 업로드 API 호출:', uploadRequest);
-      
-      // API 호출
-      const response = await storeSummaryApi.uploadCsvData(uploadRequest);
-      
-      if (response.success) {
-        console.log('✅ CSV 업로드 성공:', response.message);
-        setCsvUploadSuccess(true);
-        // sales 항목을 완료로 표시
-        setCompleted((prev) => ({ ...prev, sales: true }));
-        
-        // 파일명 저장
-        setFileNames((prev) => ({ ...prev, sales: file.name }));
-        
-        // 로컬 스토리지에 상태 저장
-        try {
-          localStorage.setItem('hybridStart.completed', JSON.stringify({ ...completed, sales: true }));
-          localStorage.setItem('hybridStart.files', JSON.stringify({ ...fileNames, sales: file.name }));
-        } catch (storageError) {
-          console.warn('로컬 스토리지 저장 실패:', storageError);
-        }
-      } else {
-        throw new Error(response.message || 'CSV 업로드에 실패했습니다.');
-      }
-      
-    } catch (error: any) {
-      console.error('❌ CSV 업로드 실패 - 전체 에러 객체:', error);
-      console.error('❌ 에러 응답:', error?.response);
-      console.error('❌ 에러 응답 데이터:', error?.response?.data);
-      console.error('❌ 에러 상태 코드:', error?.response?.status);
-      console.error('❌ 에러 메시지:', error?.message);
-      console.error('❌ 요청 설정:', error?.config);
-      
-      // 서버 에러 응답에서 더 구체적인 메시지 추출
-      let errorMessage = 'CSV 업로드 중 오류가 발생했습니다.';
-      
-      if (error?.response?.status === 500) {
-        const serverError = error?.response?.data?.message || error?.response?.data?.error || 'Internal Server Error';
-        errorMessage = `서버 오류 (500): ${serverError}`;
-        console.error('❌ 서버 500 오류 상세:', serverError);
-      } else if (error?.response?.status === 400) {
-        errorMessage = error?.response?.data?.message || '잘못된 요청입니다. CSV 파일 형식을 확인해주세요.';
-      } else if (error?.response?.status === 404) {
-        errorMessage = '요청한 리소스를 찾을 수 없습니다.';
-      } else if (error?.response?.data?.message) {
-        errorMessage = error.response.data.message;
-      } else if (error?.message && error.message !== '[object Object]') {
-        errorMessage = error.message;
-      } else {
-        errorMessage = `네트워크 오류가 발생했습니다. (상태: ${error?.response?.status || 'unknown'})`;
-      }
-      
-      setCsvUploadError(errorMessage);
-    } finally {
-      setIsUploadingCsv(false);
-    }
-  };
 
   const handleSubmit = async () => {
     setShowResult(false);
@@ -506,6 +329,7 @@ const StartHybridEvaluation = () => {
             completed={completed}
             error={error}
             isSubmitting={isSubmitting}
+            sessionId={sessionId || ''}
             onHelpClick={() => setHelpModalOpen(true)}
             onAttachHelpClick={() => setAttachHelpModalOpen(true)}
             onFileUpload={(f) => {
@@ -526,6 +350,21 @@ const StartHybridEvaluation = () => {
                   JSON.stringify(nextCompleted),
                 );
               } catch {}
+            }}
+            onCsvUploadComplete={(fileName) => {
+              // sales 항목을 완료로 표시
+              setCompleted((prev) => ({ ...prev, sales: true }));
+              
+              // 파일명 저장
+              setFileNames((prev) => ({ ...prev, sales: fileName }));
+              
+              // 로컬 스토리지에 상태 저장
+              try {
+                localStorage.setItem('hybridStart.completed', JSON.stringify({ ...completed, sales: true }));
+                localStorage.setItem('hybridStart.files', JSON.stringify({ ...fileNames, sales: fileName }));
+              } catch (storageError) {
+                console.warn('로컬 스토리지 저장 실패:', storageError);
+              }
             }}
             onSubmit={handleSubmit}
           />
